chore(deps): drop the typescript compiler from runtime deps (closes #402, closes #421)
typescript becomes a devDependency (the tsc build/type-check tool), so it no longer
ships to users (npm install --omit=dev): −8.4 MB bundle / −23 MB node_modules.

The last runtime consumers were the dispatch fact-rules:
- Dropped low-signal style smells: commented-out-code, duplicate-string-literal,
  max-switch-cases, no-magic-numbers, no-boolean-params, no-complex-conditionals.
- no-magic-numbers + high-entropy-string were already dead (never registered;
  secrets are covered by the gitleaks extractor).
- dynamic-regexp/ReDoS dropped in favor of the pattern-based ast-grep
  redos-nested-quantifier + tree-sitter unsafe-regex rules.
- function-in-loop → follow-up ast-grep rule (#428).
- Survivors (cors-wildcard, no-commented-credentials, high-import-coupling) are now
  one-file-per-rule under dispatch/rules/ and register EAGERLY in integration.ts.

Unifies fact-rule registration (#421) and removes the obsolete lazy
ensureTypeScriptDispatchUnits degrade indirection from fact-runner: with no
typescript in the graph and web-tree-sitter loaded via dynamic import, the
tree-sitter providers register eagerly too. Deleted the deps/typescript accessor
+ the now-moot eager-typescript guard, replaced by a stronger runtime-free guard
(no shipped source imports typescript). deps-centralization list updated.

FP check on the survivors over the codebase: cors-wildcard 0, no-commented-
credentials 0, high-import-coupling 11 (all true positives — genuinely wide files).
build + lint + full suite (2783) green.
