+++ package.json
@@ -85,7 +85,6 @@
 		"js-yaml": "^4.1.1",
 		"minimatch": "^10.2.5",
 		"typebox": "^1.0.0",
-		"typescript": "^6.0.3",
 		"vscode-jsonrpc": "^9.0.0"
 	},
 	"peerDependencies": {
@@ -105,6 +104,7 @@
 		"@types/js-yaml": "^4.0.9",
 		"@types/node": "^26.0.0",
 		"@vitest/coverage-v8": "^4.1.5",
+		"typescript": "^6.0.3",
 		"typescript-language-server": "^5.1.3",
 		"vitest": "^4.1.1"
 	},
